// app/(dashboard)/packages/types.ts
export interface DeliverableWithQuantity {
  deliverableId: string;
  quantity: string;
}

export interface Package {
  id?: string;
  name: string;
  price: number;
  eventId: string;
  shoots: string[];
  deliverables: DeliverableWithQuantity[];
  createdAt?: Date | { seconds: number } | string;
}